bfs and still fixing

+++ P-Emerald_2/arm9/source/map2d.cpp
@@ -1,34 +1,34 @@
 /*
-    Pokémon Emerald 2 Version
-    ------------------------------
+Pokémon Emerald 2 Version
+------------------------------
 
-    p_file        : map2d.cpp
-    author      : Philip Wellnitz (RedArceus)
-    description : Main engine for parsing and handling 2D maps.
+p_file        : map2d.cpp
+author      : Philip Wellnitz (RedArceus)
+description : Main engine for parsing and handling 2D maps.
 
-    Copyright (C) 2012 - 2014
-    Philip Wellnitz (RedArceus)
+Copyright (C) 2012 - 2014
+Philip Wellnitz (RedArceus)
 
-    This software is provided 'as-is', without any express or implied
-    warranty.  In no event will the authors be held liable for any
-    damages arising from the use of this software.
+This software is provided 'as-is', without any express or implied
+warranty.  In no event will the authors be held liable for any
+damages arising from the use of this software.
 
-    Permission is granted to anyone to use this software for any
-    purpose, including commercial applications, and to alter it and
-    redistribute it freely, subject to the following restrictions:
+Permission is granted to anyone to use this software for any
+purpose, including commercial applications, and to alter it and
+redistribute it freely, subject to the following restrictions:
 
 
-    1.	The origin of this software must not be misrepresented; you
-    must not claim that you wrote the original software. If you use
-    this software in a product, an acknowledgment in the product
-    is required.
+1.	The origin of this software must not be misrepresented; you
+must not claim that you wrote the original software. If you use
+this software in a product, an acknowledgment in the product
+is required.
 
-    2.	Altered source versions must be plainly marked as such, and
-    must not be misrepresented as being the original software.
+2.	Altered source versions must be plainly marked as such, and
+must not be misrepresented as being the original software.
 
-    3.	This notice may not be removed or altered from any source
-    distribution.
-    */
+3.	This notice may not be removed or altered from any source
+distribution.
+*/
 
 
 #include "map2d.h"
@@ -59,26 +59,26 @@ namespace map2d {
             fread( p_palette[ i ].m_pal, sizeof( u16 ) * 16, 1, p_file );
         fclose( p_file );
     }
-    inline void readTileSet( FILE* p_file, TileSet& p_tileSet, u16 p_startIdx = 0, u16 p_size = 512 ) {
+    inline void readTileSet( FILE* p_file, TileSet& p_tileSet, int p_startIdx = 0, int p_size = 512 ) {
         if( p_file == 0 )
             return;
         fread( &p_tileSet.m_blocks[ p_startIdx ], sizeof( Tile )*p_size, 1, p_file );
         fclose( p_file );
     }
-    inline void readNop( FILE* p_file, u16 p_cnt ) {
+    inline void readNop( FILE* p_file, int p_cnt ) {
         if( p_file == 0 )
             return;
         fread( 0, sizeof( u8 )*p_cnt, 1, p_file );
     }
-    inline void readBlockSet( FILE* p_file, BlockSet& p_tileSet, u16 p_startIdx = 0, u16 p_size = 512 ) {
+    inline void readBlockSet( FILE* p_file, BlockSet& p_tileSet, int p_startIdx = 0, int p_size = 512 ) {
         if( p_file == 0 )
             return;
         readNop( p_file, 4 );
-        for( u16 i = 0; i < p_size; ++i ) {
+        for( int i = 0; i < p_size; ++i ) {
             fread( p_tileSet.m_blocks[ p_startIdx + i ].m_bottom, 4 * sizeof( BlockAtom ), 1, p_file );
             fread( p_tileSet.m_blocks[ p_startIdx + i ].m_top, 4 * sizeof( BlockAtom ), 1, p_file );
         }
-        for( u16 i = 0; i < p_size; ++i ) {
+        for( int i = 0; i < p_size; ++i ) {
             fread( &p_tileSet.m_blocks[ p_startIdx + i ].m_bottombehave, sizeof( u8 ), 1, p_file );
             fread( &p_tileSet.m_blocks[ p_startIdx + i ].m_topbehave, sizeof( u8 ), 1, p_file );
         }
@@ -97,7 +97,7 @@ namespace map2d {
             fread( &a.m_maxFrame, sizeof( u8 ), 1, p_file );
             a.m_acFrame = 0;
             a.m_animationTiles.assign( a.m_maxFrame, Tile( ) );
-            for( u16 i = 0; i < a.m_maxFrame; ++i )
+            for( int i = 0; i < a.m_maxFrame; ++i )
                 fread( &a.m_animationTiles[ i ], sizeof( Tile ), 1, p_file );
             p_animations.push_back( a );
         }
@@ -152,10 +152,10 @@ namespace map2d {
         sprintf( buffer, "%s%s.anb", p_path, p_name );
         FILE* A = fopen( buffer, "r" );
         if( A ) {
-            u16 N; fscanf( A, "%d", &N );
+            int N; fscanf( A, "%d", &N );
             m_anbindungen = std::vector<Anbindung>( N );
             //printf("TEST3 : %i\n",N);
-            for( u16 i = 0; i < N; ++i ) {
+            for( int i = 0; i < N; ++i ) {
                 Anbindung& ac = m_anbindungen[ i ];
                 fscanf( A, "%s %c %d %d", ac.m_name, &ac.m_direction, &ac.m_move, &ac.m_mapidx );
                 //printf("TEST4 : %s %c %d %d\n",ac.name, ac.p_direction,ac.move,ac.mapidx);
@@ -170,8 +170,8 @@ namespace map2d {
                 readNop( mapF2, 12 );
 
                 //TODO: SPEED-UP this Part
-                for( u16 x = 0; x < ac.m_mapsx; ++x )
-                    for( u16 y = 0; y < ac.m_mapsy; ++y ) {
+                for( int x = 0; x < ac.m_mapsx; ++x )
+                    for( int y = 0; y < ac.m_mapsy; ++y ) {
                     if( ac.m_direction == 'W' && y >= ac.m_mapsy - 10 && x + ac.m_move + 10 >= 0 && x + ac.m_move < (int)m_sizex + 10 )
                         fread( &( m_blocks[ x + ac.m_move + 10 ][ y - ac.m_mapsy + 10 ] ), sizeof( MapBlockAtom ), 1, mapF2 );
 
@@ -190,33 +190,33 @@ namespace map2d {
                 fclose( mapF2 );
             }
             consoleSelect( &Bottom );
-            for( u8 i = 0; i < 2; ++i )
-                for( u8 j = 0; j < 2; ++j ) {
+            for( int i = 0; i < 2; ++i )
+                for( int j = 0; j < 2; ++j ) {
                 if( fscanf( A, "%hu", &m_rand[ i ][ j ] ) == EOF )
                     m_rand[ i ][ j ] = 0;
                 }
             fclose( A );
         }
 
-        for( u16 x = 0; x < (int)m_sizex; ++x )
+        for( int x = 0; x < (int)m_sizex; ++x )
             fread( &m_blocks[ x + 10 ][ 10 ], sizeof( MapBlockAtom )*m_sizey, 1, mapF );
         //printf("Test5 %i",anbindungen.p_size());
         fclose( mapF );
     }
 
-    u16 tcnt = 0;
-    u16 bgs[ 4 ];
+    int tcnt = 0;
+    int bgs[ 4 ];
 
-    u16 lastrow, lastcol;
-    u16 lastbx, lastby;
+    int lastrow, lastcol;
+    int lastbx, lastby;
 
-    void Map::fill( u16* p_mapMemory[ 4 ], u16 p_xmin, u16 p_x, u16 p_xmax, u16 p_ymin, u16 p_y, u16 p_ymax, u16 p_c ) {
-        u16 toplayer = 1, betw = 2, bottomlayer = 3;
+    void Map::fill( u16* p_mapMemory[ 4 ], int p_xmin, int p_x, int p_xmax, int p_ymin, int p_y, int p_ymax, int p_c ) {
+        int toplayer = 1, betw = 2, bottomlayer = 3;
 
         Block acBlock;
-        if( p_x < 0 || p_y < 0 || p_x >= m_sizex + 20 || p_y >= m_sizey + 20 )
+        if( p_x < 0 || p_y < 0 || p_x >= (int)m_sizex + 20 || p_y >= (int)m_sizey + 20 )
             acBlock = m_blockSets.m_blocks[ m_rand[ p_x % 2 ][ p_y % 2 ] ];
-        else if( ( p_x < 10 || p_y < 10 || p_x >= m_sizex + 10 || p_y >= m_sizey + 10 ) && m_blocks[ p_x ][ p_y ].m_blockidx == 0 )
+        else if( ( p_x < 10 || p_y < 10 || p_x >= (int)m_sizex + 10 || p_y >= (int)m_sizey + 10 ) && m_blocks[ p_x ][ p_y ].m_blockidx == 0 )
             acBlock = m_blockSets.m_blocks[ m_rand[ p_x % 2 ][ p_y % 2 ] ];
         else
             acBlock = m_blockSets.m_blocks[ m_blocks[ p_x ][ p_y ].m_blockidx ];
@@ -246,19 +246,19 @@ namespace map2d {
         }
     }
 
-    void Map::movePlayer( u16 p_direction ) {
-        u16 c = 0;
+    void Map::movePlayer( int p_direction ) {
+        int c = 0;
         u16* mapMemory[ 4 ];
-        for( u8 i = 1; i < 4; ++i )  mapMemory[ i ] = (u16*)BG_MAP_RAM( 2 * i - 1 );
+        for( int i = 1; i < 4; ++i )  mapMemory[ i ] = (u16*)BG_MAP_RAM( 2 * i - 1 );
 
-        u16 bx = lastbx, by = lastby;
+        int bx = lastbx, by = lastby;
 
-        u16 xmin = 0,
+        int xmin = 0,
             xmax = 0,
             ymin = 0,
             ymax = 0,
             plsval = 0;
-        u16 lc = lastcol, lr = lastrow;
+        int lc = lastcol, lr = lastrow;
 
         switch( p_direction ) {
             case 1:
@@ -314,15 +314,15 @@ namespace map2d {
                 plsval = 32;
 
         }
-        u16 c2 = c;
+        int c2 = c;
         if( p_direction == 1 ) {
             ymin += 16;
             ymax += 16;
         }
-        for( u8 i = 0; i < 2; ++i ) {
+        for( int i = 0; i < 2; ++i ) {
             if( ( p_direction % 2 == 0 ) || ( i == ( ( lastrow + ( p_direction == 3 ? 1 : 0 ) ) % 32 ) / 16 ) )
-                for( u16 x = xmin; x < xmax; x++ ) {
-                for( u16 y = ymin; y < ymax; y++ ) {
+                for( int x = xmin; x < xmax; x++ ) {
+                for( int y = ymin; y < ymax; y++ ) {
 
                     fill( mapMemory, xmin, x, xmax, ymin, y, ymax, c );
 
@@ -340,53 +340,53 @@ namespace map2d {
         }
         switch( p_direction ) {
             case 2: case 4:
-                for( u8 i = 1; i < 4; ++i ) {
+                for( int i = 1; i < 4; ++i ) {
                     u16 q[ 64 ];
-                    for( u8 g = 0; g < 32; ++g )
+                    for( int g = 0; g < 32; ++g )
                         q[ g ] = mapMemory[ i ][ c2 + g ];
-                    for( u8 g = 0; g < 32; ++g )
+                    for( int g = 0; g < 32; ++g )
                         q[ g + 32 ] = mapMemory[ i ][ c2 + g + 1024 ];
-                    for( u8 o = 0; o < 32; ++o )
+                    for( int o = 0; o < 32; ++o )
                         mapMemory[ i ][ c2 + o ] = q[ ( o + 64 - 2 * ( ( lr + 1 ) % 32 ) ) % 64 ];
-                    for( u8 o = 0; o < 32; ++o )
+                    for( int o = 0; o < 32; ++o )
                         mapMemory[ i ][ c2 + o + 1024 ] = q[ ( o + 96 - 2 * ( ( lr + 1 ) % 32 ) ) % 64 ];
                 }
                 c2 += 32;
-                for( u8 i = 1; i < 4; ++i ) {
+                for( int i = 1; i < 4; ++i ) {
                     u16 q[ 64 ];
-                    for( u8 g = 0; g < 32; ++g )
+                    for( int g = 0; g < 32; ++g )
                         q[ g ] = mapMemory[ i ][ c2 + g ];
-                    for( u8 g = 0; g < 32; ++g )
+                    for( int g = 0; g < 32; ++g )
                         q[ g + 32 ] = mapMemory[ i ][ c2 + g + 1024 ];
-                    for( u8 o = 0; o < 32; ++o )
+                    for( int o = 0; o < 32; ++o )
                         mapMemory[ i ][ c2 + o ] = q[ ( o + 64 - 2 * ( ( lr + 1 ) % 32 ) ) % 64 ];
-                    for( u8 o = 0; o < 32; ++o )
+                    for( int o = 0; o < 32; ++o )
                         mapMemory[ i ][ c2 + o + 1024 ] = q[ ( o + 96 - 2 * ( ( lr + 1 ) % 32 ) ) % 64 ];
                 }
                 break;
             case 1: case 3:
                 c2 += ( ( ( lastrow + ( p_direction == 3 ? 1 : 0 ) ) % 32 ) / 16 ) * ( 1024 - 32 );
 
-                for( u8 i = 1; i < 4; ++i ) {
+                for( int i = 1; i < 4; ++i ) {
                     u16 q[ 32 ];
-                    for( u8 g = 0; g < 32; ++g )
+                    for( int g = 0; g < 32; ++g )
                         q[ g ] = mapMemory[ i ][ c2 + 32 * g ];
-                    for( u8 o = 0; o < 32; ++o )
+                    for( int o = 0; o < 32; ++o )
                         mapMemory[ i ][ c2 + 32 * o ] = q[ ( o + 32 - 2 * ( lc + 1 ) ) % 32 ];
                 }
                 c2++;
-                for( u8 i = 1; i < 4; ++i ) {
+                for( int i = 1; i < 4; ++i ) {
                     u16 q[ 32 ];
-                    for( u8 g = 0; g < 32; ++g )
+                    for( int g = 0; g < 32; ++g )
                         q[ g ] = mapMemory[ i ][ c2 + 32 * g ];
-                    for( u8 o = 0; o < 32; ++o )
+                    for( int o = 0; o < 32; ++o )
                         mapMemory[ i ][ c2 + 32 * o ] = q[ ( o + 32 - 2 * ( lc + 1 ) ) % 32 ];
                 }
                 break;
         }
     }
 
-    void Map::draw( u16 p_bx, u16 p_by, bool p_init ) {
+    void Map::draw( int p_bx, int p_by, bool p_init ) {
         if( p_init ) {
             videoSetMode( MODE_0_2D /*| DISPLAY_BG0_ACTIVE*/ | DISPLAY_BG1_ACTIVE |
                           DISPLAY_BG2_ACTIVE | DISPLAY_BG3_ACTIVE | DISPLAY_SPR_ACTIVE | DISPLAY_SPR_1D );
@@ -397,7 +397,7 @@ namespace map2d {
             //REG_BG2CNT = BG_32x32 | BG_COLOR_16 |   BG_MAP_BASE(2) | BG_TILE_BASE(1)| BG_PRIORITY(2);
             //REG_BG3CNT = BG_32x32 | BG_COLOR_16 |   BG_MAP_BASE(3) | BG_TILE_BASE(1)| BG_PRIORITY(3);
 
-            for( u8 i = 1; i < 4; ++i ) {
+            for( int i = 1; i < 4; ++i ) {
                 bgs[ i ] = bgInit( i, BgType_Text4bpp, BgSize_T_512x256, 2 * i - 1, 1 );
                 bgSetPriority( bgs[ i ], i );
                 bgScroll( bgs[ i ], 128, 32 );
@@ -407,13 +407,13 @@ namespace map2d {
 
             u8* tileMemory = (u8*)BG_TILE_RAM( 1 );
 
-            for( u16 i = 0; i < 1024; ++i )
+            for( int i = 0; i < 1024; ++i )
                 swiCopy( m_tileset.m_blocks[ i ].m_tile, tileMemory + i * 32, 16 );
 
             dmaCopy( m_pals, BG_PALETTE, 512 );
         }
         u16* mapMemory[ 4 ];
-        for( u8 i = 1; i < 4; ++i )  mapMemory[ i ] = (u16*)BG_MAP_RAM( 2 * i - 1 );
+        for( int i = 1; i < 4; ++i )  mapMemory[ i ] = (u16*)BG_MAP_RAM( 2 * i - 1 );
 
         lastrow = 31;
         lastcol = 15;
@@ -425,10 +425,10 @@ namespace map2d {
 
         lastbx = p_bx;
         lastby = p_by;
-        for( u8 i = 0; i < 2; ++i ) {
-            for( u16 x = p_by; x < p_by + 17; x++ ) {
-                for( u16 y = p_bx; y < p_bx + 17; y++ ) {
-                    u16 toplayer = 1, bottomlayer = 3;
+        for( int i = 0; i < 2; ++i ) {
+            for( int x = p_by; x < p_by + 17; x++ ) {
+                for( int y = p_bx; y < p_bx + 17; y++ ) {
+                    int toplayer = 1, bottomlayer = 3;
 
                     Block acBlock = m_blockSets.m_blocks[ m_blocks[ x ][ y ].m_blockidx ];
 
@@ -436,9 +436,9 @@ namespace map2d {
                     //    acBlock = b.blocks[tcnt];
                     //    tcnt = (tcnt +1)%1024;
                     //}
-                    if( x < 0 || y < 0 || x >= m_sizex + 20 || y >= m_sizey + 20 ) {
+                    if( x < 0 || y < 0 || x >= (int)m_sizex + 20 || y >= (int)m_sizey + 20 ) {
                         acBlock = m_blockSets.m_blocks[ m_rand[ x % 2 ][ y % 2 ] ];
-                    } else if( ( x < 10 || y < 10 || x >= m_sizex + 10 || y >= m_sizey + 10 ) && m_blocks[ x ][ y ].m_blockidx == 0 ) {
+                    } else if( ( x < 10 || y < 10 || x >= (int)m_sizex + 10 || y >= (int)m_sizey + 10 ) && m_blocks[ x ][ y ].m_blockidx == 0 ) {
                         acBlock = m_blockSets.m_blocks[ m_rand[ x % 2 ][ y % 2 ] ];
                     } else {
                         //consoleSelect(&Top);
